keep operating mode for Z1/Z2 in sync

+++ components/ecodan/ecodan.h
@@ -48,7 +48,7 @@ namespace ecodan
         void set_dhw_force(bool on);
         void set_holiday(bool on);
         void set_power_mode(bool on);
-        void set_hp_mode(int mode);
+        void set_hp_mode(uint8_t mode);
         void set_controller_mode(CONTROLLER_FLAG flag, bool on);
         void set_uart_parent(uart::UARTComponent *uart) { this->uart_ = uart; }
         void set_proxy_uart(uart::UARTComponent *uart) { this->proxy_uart_ = uart; }
